Reorganize folders: bots/ → football/bots/, maccabipedia_calendar/ → calendar/, delete old/
- Move bots/ into football/bots/ (football-specific, not generic)
- Rename maccabipedia_calendar/ → calendar/ (remove redundant prefix)
- Delete deprecated old/ folder (Walla scripts, unused)
- Update all GitHub Actions workflows: use pip install -e . and python -m module paths
