+++ .github/workflows/update_maccabipedia_volleyball_calendar.yaml
@@ -14,7 +14,7 @@ jobs:
           python-version: 3.11
           architecture: x64
       - name: Install dependencies
-        run: pip install -r requirements.txt
+        run: pip install -e .
       - name: Update MaccabiPedia Calendar
         env:
           VOLLEYBALL_CALENDAR_ID: ${{ secrets.VOLLEYBALL_CALENDAR_ID }}
@@ -22,7 +22,7 @@ jobs:
           PYWIKIBOT_DIR: pywikibot_configs/
           MACCABIPEDIA_BOT_USERNAME: ${{ secrets.MACCABIPEDIA_BOT_USERNAME }}
           MACCABIPEDIA_BOT_PASSWORD: ${{ secrets.MACCABIPEDIA_BOT_PASSWORD }}
-        run: python maccabipedia_calendar/main_volleyball.py
+        run: python -m maccabipediabot.calendar.main_volleyball
   workflow-keepalive:
     if: github.event_name == 'schedule'
     runs-on: ubuntu-latest
