--- requirements.txt
@@ -1,14 +1,0 @@
-pytest>=8.0
-pydantic>=2.0
-mwparserfromhell==0.6.6
-pywikibot==9.6.0
-requests==2.32.3
-pandas==2.2.3
-lxml<5,>=4.9.1
-maccabistats>=2.50
-html5lib==1.1
-google-api-python-client==1.12.8
-google-auth-oauthlib==0.4.2
-beautifulsoup4>=4.9.3
-python-dotenv==0.15.0
-tldextract==5.3.0
