+++ src/maccabipediabot/calendar/VOLLEYBALL_MANUAL_GAMES.md
