+++ src/maccabipediabot/calendar/main_volleyball.py
@@ -11,8 +11,8 @@ from typing import Dict, List
 import requests
 from dotenv import load_dotenv
 
-from maccabipediabot.maccabipedia_calendar.calendar_operations import fetch_games_from_calendar, update_event, upload_event, delete_event, Event
-from maccabipediabot.maccabipedia_calendar.google_calendar_api import initialize_global_google_service_account_from_memory_json
+from maccabipediabot.calendar.calendar_operations import fetch_games_from_calendar, update_event, upload_event, delete_event, Event
+from maccabipediabot.calendar.google_calendar_api import initialize_global_google_service_account_from_memory_json
 from maccabipediabot.volleyball.upload_volleyball_games_from_iva_site import extract_games_metadata
 from maccabipediabot.volleyball.volleyball_game import VolleyballGame
